requirements.txt
torchvision==0.5.0
pretrainedmodels==0.7.4
numpy==1.16.4
scipy==1.5.2
tqdm==4.32.2
torch==1.4.0
Pillow==8.1.0
scikit_learn==0.24.0
